Add files via upload
Build the general architecture of a learning algorithm, including Initializing parameters, Calculating the cost function and its gradient, Using an optimization algorithm (gradient descent), Gather all three functions above into the main model function, in the right order.

The general architecture of a learning algorithm was constructed to learn and recognize images of cats. Parameters are initialized, the cost function and its gradient is calculated, an optimization algorithm (gradient descent) is implemented. All three functions are then gathered into the main model so it trains on a sample of over 200 pre-processed cat images. The model is then tested on a dataset of 50 images. A test accuracy of 68% was achieved.